Add files via upload
Test Data and Stock Data files for HW02 with Macros are uploaded here. I encountered a LOT of trouble running the Macro to determine the volume for each stock which I separated out into a separate module in an attempt to keep Excel from freezing. I got the volume calculated on the test data, but was not able to get the same macro to produce results on the real data with Excel freezing and me having to force quit. Looked into several causes for this, but working through some of the potential issues did not improve performance.